Transaction Class Implementation
- add transaction logging to methods in Account class

- test file creation

+++ src/main/java/Model.java
@@ -1,10 +1,14 @@
+import com.sun.tools.javac.code.Attribute;
 import de.uniks.networkparser.graph.Cardinality;
 import de.uniks.networkparser.graph.Clazz;
 import de.uniks.networkparser.graph.DataType;
+import de.uniks.networkparser.graph.Parameter;
 import org.sdmlib.models.classes.ClassModel;
 import org.sdmlib.storyboards.Storyboard;
+import swe443.bluebank.Transaction;
 
 import java.math.BigInteger;
+import java.util.Enumeration;
 
 
 /**
@@ -47,7 +51,7 @@ public class Model {
                 .withMethod("viewBalance", DataType.VOID);
 
         Clazz transactionClass = model.createClazz("Transaction")
-                .withAttribute("amount", DataType.INT)
+                .withAttribute("amount", DataType.DOUBLE)
                 .withMethod("logbuilder",DataType.STRING);
